#8, #9: Refactoring to improve route security (logged in user can only update her cart)
