+++ server/api/users.js
@@ -12,21 +12,3 @@ router.get('/', (req, res, next) => {
     .then(users => res.json(users))
     .catch(next)
 })
-
-router.put('/:id', async (req, res, next) => {
-  if (req.user.id === req.params.id) {
-    const userInfo = {
-      cart: req.body.cart,
-    }
-    await User.update(userInfo, {
-      where: {
-        id: req.params.id
-      }
-    })
-    res.sendStatus(200)
-  } else {
-    const error = new Error('Not Authorized')
-    error.status = 401
-    next(error)
-  }
-})
